fixed several cases in which size_t was compared to unsigned int - 64-bit
Former-commit-id: 1fe95b7f0411b520522306c4d32336827db1b20b

+++ branches/ocelot-ptx-2.1/ocelot/executive/implementation/EmulatorCallStack.cpp
@@ -182,7 +182,7 @@ namespace executive
 		*((unsigned int*) &_stack[
 			_stackPointer + 2 * sizeof(unsigned int)]) = callerFrameSize;	
 		
-		_sharedMemory.resize(std::max(_sharedMemory.size(), sharedSize));
+		_sharedMemory.resize(std::max(_sharedMemory.size(), (size_t)sharedSize));
 	}
 
 	void EmulatorCallStack::popFrame()
